nexd: Make Informer return a map of devices
Previously, the Informer returned a device list. Now it returns it in
the form of a map of public keys to devices. This is needed for future
changes that will be doing cross referencing of peer information from
two difference sources - the Nexodus API and local wireguard
configuration. The public key is the common unique identifier used in
both places.

Signed-off-by: Russell Bryant <rbryant@redhat.com>