nexd: Change key for the device cache
Change the device cache to be keyed on the device public key instead
of Id. The public key is the common piece of identifying information
available across multiple sources of information.

Also track a timestamp for device cache entries. This will be used in
future changes when we need to calculate how long we've had the latest
device configuration applied.

Signed-off-by: Russell Bryant <rbryant@redhat.com>