nexd: Store peer configs in a map
This is another case where we have a set of peers we need to cross
reference with another set of peer info. Accessing it by public key
will be faster than linear searches for larger organizations.

There should be no logic changes in this patch, only a change in the
storage. Benefits of the new storage will come in future patches as
more code is introduced that needs to reference the last known config.

Signed-off-by: Russell Bryant <rbryant@redhat.com>