nexd: Choose a default organization
A recent change added support in nexd for users that are members of
multiple organizations. This change restores the original simple form
of running nexd: `nexd URL`.

If a user is a member of multiple organizations and no org ID was
specified, choose the default organization created for the user, which
is the one that has the same name as the user.

In the future we may want to add support in the API and UI for
specifying which organization is the default, but this seems like a
good enough start, and it just keeps the same behavior from before
when only 1 org (the user's org) was supported.

Closes #1094

Signed-off-by: Russell Bryant <rbryant@redhat.com>