docs: Add nexctl usage docs
Update the `nexctl` docs with usage information based on the help
text. This doc can be automatically updated using the
`contrib/nexctl-docs.sh` script.

There is plenty of cleanup or additions that could be made. If the
help text is improved or expanded, that will make its way to the doc
site as well by running this script.

Closes #1049

Signed-off-by: Russell Bryant <rbryant@redhat.com>