nexd: Allow nexd to change peer configs
This change should introduce no functional changes, but it reorganizes
the code in an important way.

Previously, the only time `nexd` would reconfigure its peers is when
it saw a configuration change come in for that peer from the API. The
logic is now changed such that the peer will be configured anytime
that nexd's peer configuration calculations decide. It may change
based on a config change from the API, or it may change because `nexd`
decides it should change for some other reason.

A future change will take advantage of this shift to start making peer
configuration changes based on nexd's own observations on the
connection health to that peer.

Signed-off-by: Russell Bryant <russell.bryant@gmail.com>