mergify: Drop use of changes-requested label
Remove the rules that add/remove the changes-requested label.

This does not behave as I expected. It adds the label when I wanted,
but it does not remove it. I expected the label to get removed as soon
as someone re-requests review after changes had been requested.
Instead, that state doesn't change until that reviewer approves the
PR. Based on this behavior, it's not a state we can rely on to signal
when a PR is ready to be reviewed again.

Signed-off-by: Russell Bryant <russell.bryant@gmail.com>