rpm: Fix building srpm in aarch64 rpm jobs
When building an srpm to submit to copr for an aarch64 build, just
build the srpm in an x86_64 build root. The arch doesn't matter when
building the srpm, and it'll be more efficient as it will match the
architecture of the host building the srpm.

When setting up the build root for aarch64 for use with `make rpm`,
install `qemu-user-static-aarch64`, as this is needed to build the
aarch64 rpm on an x86_64 host.

Closes #1142

Signed-off-by: Russell Bryant <rbryant@redhat.com>