docs: Updates to network-routers docs
Fix syntax for specifying multiple child prefixes. You must specify
`--child-prefix` multiple times instead of providing a list of
prefixes to a single instance of the option.

Drop the deprecated method of specifying the service URL as a
positional argument. Simplify the docs by not including it at all.

Make various wording and grammar tweaks along the way.

Signed-off-by: Russell Bryant <rbryant@redhat.com>