Add wireguard userspace tools to the nexd
- Adds about 30m for the wg user space binary. Will be useful
to debug user issues early on. We currently get information
such as peer listings and stats directly from the wg interface
but if nexd isn't running those stats aren't available. Since
nexd does not rely on this it is removable at any time but would
like to have it available for time saving for installs for now.

Signed-off-by: Brent Salisbury <bsalisbu@redhat.com>