nexd: Restore timestamps in logs
My original reason for dropping the timestamps was two-fold: systemd
timestamps the logs for nexd and the timestamp format used in the
production config was unreadable to humans, so I was always annoyed
looking at the logs.

Relying on systemd (or something else) to timestamp the logs isn't
appropriate. When running in a container, we don't have that same
feature present.

Tweak the production log config to use a human-readable timestamp
format. The default was unix time (seconds since epoch).

Signed-off-by: Russell Bryant <rbryant@redhat.com>