nexd: Fix peering health monitoring
The code added to determine whether a healthy session between two
peers exists was flawed in multiple ways, resulting in a flapping
healthy/unhealthy state visible either in `nexctl nexd peers list`,
or in debug logs.

The wireguard paper, nexd logs, packet captures, and review of the
kernel implementation of wireguard were used to narrow this down. For
reference, here is the wireguard paper:

  https://www.wireguard.com/papers/wireguard.pdf

The wireguard kernel implementation can be found under
linux/drivers/net/wireguard in the kernel tree.

Problem 1: The check based on the last handshake time was just
completely wrong. It assumed a handshake would occur within
(persistent keepalive interval + KEEPALIVE_TIMEOUT). Not only is this
the wrong window of time to use, it's two values that are unrelated.
The keepalive timeout is a constant from the spec related to passive
keepalives and is not used with optional persistent keepalives.

The fix for problem 1 is to check that the last handshake is within
the REJECT_AFTER_TIME constant defined in the paper.

> After REJECT_AFTER_MESSAGES transport data messages or after the
> current secure session is REJECT_AFTER_TIME seconds old, whichever
> comes first, WireGuard will refuse to send or receive any more
> transport data messages using the current secure session, until a new
> secure session is created through the 1-RTT handshake.

Problem 2: The checks based on the TX and RX counters were based on
flawed assumptions about how keepalives work. The code assumed that we
should always see the TX and RX counters increase within (persistent
keepalive interval + KEEPALIVE_TIMEOUT). This is wrong for multiple
reasons.

First, as noted earlier, the KEEPALIVE_TIMEOUT constant is not related
to persistent keepalive handling.

Second, the code assumed that when persistent keepalives are enabled,
they are always sent bidirectionally at that interval. In practice, my
observation in testing is that happens ... sometimes. There are two
sources of reading helpful in understanding the observed behavior.

One source, of course, is the Wireguard paper. The protocol details
are important, but also that it describes a design goal of being
quiet and minimizing the number of messages necessary to setup and
maintain a session. This is helpful for security, efficiency, and
stealth.

However, the paper only makes one passing mention of persistent
keepalives and does not define exactly how they should be handled in
the detail given for the rest of the protocol, or that was my read of
it.

Another helpful resource for me was reviewing the kernel source and
how persistent keepalives are handled. We can see that the
implementation takes steps to try to only send the keepalives if it
deems necessary. The timer for sending a persistent keepalive gets
reset anytime an authenticated message is sent OR received.

The timer reset is in wireguard/timers.c:

    /* Should be called before a packet with authentication, whether
     * keepalive, data, or handshakem is sent, or after one is received.
     */
    void wg_timers_any_authenticated_packet_traversal(struct wg_peer *peer)
    {
            if (peer->persistent_keepalive_interval)
                    mod_peer_timer(peer, &peer->timer_persistent_keepalive,
                            jiffies + peer->persistent_keepalive_interval * HZ);
    }

Next, you can see where this function is called from wireguard/send.c
and wireguard/receive.c.

On the receive side, it is called from wg_receive_handshake_packet()
and wg_packet_consume_data_done(). In other words, any time a
handshake, keepalive, or a data message is received, the timer for
SENDING a persistent keepalive gets reset.

The send side is similar. It's called from
wg_packet_send_handshake_initiation(),
wg_packet_send_handshake_response(), and wg_packet_create_data_done().
The timer for SENDING a persistent keepalive is reset anytime a
handshake initiation or response is sent, or when sending a data
packet.

The result is that persistent keepalives are not very persistent. In
my local test environment, I often observe keepalives going
bidirectionally at the same time, but it appears to be because the
timers are so close to in sync that the timers are expiring close
enough to the same time that both sides are sending a keepalive before
their timer can be reset because one was received.

If the timing is off just enough that a keepalive is received and
processed before the local timer fires, no keepalive will be sent.
This is the condition that would cause the unhealthy peer status
flapping in the nexd logs.

Once in the state where nexd believed the session was unhealthy, the
most reliable point in time where you would see it restored is when
wireguard does key rotation. If not sooner, a new session will be
created approximately after REKEY_AFTER_TIME seconds. This handshake
will always bump both the TX and RX counters, so nexd would be happy
again.

After all of that, I have dropped any use of the TX and RX counters to
try to determine peer health because I don't see a way to use it to
reliably detect a broken session any faster than the calculation done
against the last handshake time.

I see two possible improvements for faster peeper failure detection:

1) Ensure that nexd always uses a customized fork of wireguard-go that
   includes a change to ensure persistent keepalives are always sent
   if no other data is being sent. The timer could be reset when sending
   other packets, but not when receiving them. That would ensure that
   bidirectional wireguard packets are flowing within a defined
   interval instead of the current case, where it can end up only
   going one direction in between key rotation.

2) Do health checking on top. ICMP is an obvious candidate, though
   it's possible for ICMP packets to be dropped by the destination
   host's firewall or some other middlebox, even though the
   wireguard session is fine, and other traffic is still working.

If we actually feel the need to do this, we should also dig into other
projects that use wireguard to see what they've come up with. In the
meantime, detection will take about 3 minutes. That's good enough for
now.

Signed-off-by: Russell Bryant <rbryant@redhat.com>