nexd: Refactor peering method choices
Define more concretely the concept of a peering method. Each method
has a name, a set of prerequisites that must be met to use it, and an
implementation of generating wireguard configuration using that
method.

There should be no functional changes other than an updated debug log
message that will now indicate which peering method is in use for a
peer.

Signed-off-by: Russell Bryant <rbryant@redhat.com>