nexd: dynamic peering
The most common scenario addressed by this change is that after
failing to peer via a peer's local IP or reflexive IP, we can fall
back to using a relay if one is available.

More generally, these changes give a chosen peering method a chance to
work and then fall back to alternative methods. For more examples:

- An attempt to peer via local IP will fall back to peering via
  reflexive IP

- An attempt to peer via reflexive IP will fall back to using a relay.

- If there is not currently a healthy relay available, it will fall
  back to the beginning and choose the first available method to try

These retries will keep going until something works, or it will stay
with the relay if one is available. When using a relay, other methods
will be retried after the API sends us changes to a device that
impact peering configuration.

All of this is based on peering health inferred via wireguard stats,
not testing connectivity to each device via ICMP or something else.

Closes #965
Closes #1322

Signed-off-by: Russell Bryant <rbryant@redhat.com>