Aliased predefined ports in sec group rules edit UI
- The predefined ports were getting passed as-is instead of being
normalized to a proper protocol, e.g., MYSQL needs to be mapped to
TCP to be punched into netfilter. The secgroup rules validation patch
brought this to the surface. The ports do not grey out as they did,
but I added a todo to do it in another patch to avoid regression here.

Signed-off-by: Brent Salisbury <bsalisbu@redhat.com>