rpm: don't overrwrite config on upgrade
The nexodus service config was not properly marked as a config file in
the rpm spec file. This resulted in overwriting the config every time
the package was upgraded. With this change, if the package includes an
updated version of the default configuration is put in a
`nexodus.rpmnew` file.

In passing, the git hash used in the version is now explicitly limited
to 6 characters. It's usually 6, but when I got a longer one, it broke
my rpm build.

Closes #1495

Signed-off-by: Russell Bryant <rbryant@redhat.com>