apiserver: improve the registration token apis.
The registration and device bearer tokens are now shorter secrets to improve the UX.
The short secrets are looked in the DB and converted to JWT per request (this also avoid needing to deal with expiring JWTs since the JWTs are not given to the clients).
To be even more sure we don’t give out a device token accidentally to the wrong person, we now encrypt it using the device’s public WG key.

Signed-off-by: Hiram Chirino <hiram@hiramchirino.com>