e2e: Allow for some packet drops in udp tests
This e2e test is one that we commonly see some failures from. For
example, the  last one I saw showed 4 responses received via UDP
instead of 5. Since it's UDP, and we're running these tests in
parallel so load could be an issue at times, allow for some drops as
long as we receive the majority of responses we expected.

Signed-off-by: Russell Bryant <rbryant@redhat.com>