nexd: Fix delay of peering fallback
The code intended to delay dropping back to the next preferred peering
method if a given method was previously working and then stopped. The
intent was to wait 3 minutes in that case, instead of waiting much
less time for a method that had not yet worked at all.

The code was checking the wrong timestamp for this. The problem was
discovered while writing unit tests for this code.

Signed-off-by: Russell Bryant <rbryant@redhat.com>