nexd: Fix peering reset when no relays are present
The code that intended to make peering start over at the beginning
would only work after connecting with a relay and then that relay goes
away. If no relay was available, we would never go back to the
beginning of the peering method list.

This change also introduces a "none" peering method to be more
explicit when we do not think it is currently possible to reach a
peer via any means. Previously, "via-relay" was used for both
"reachable by a relay" and "reachable IF a relay was present".

Signed-off-by: Russell Bryant <rbryant@redhat.com>