nexd: Remove --stun option
This option appears to be left over from a time when we only kept
track of one endpoint address per device. We would either just use a
local address or we would use one from stun. We now intentionally keep
track of both in the Endpoints list.

The code for `--stun` now did something that didn't make any sense. If
`--stun` is used, it stores the reflexive IP from stun in the "local"
endpoint as well as the stun endpoint.

As far as I can tell, there is no longer any value to this option. We
are always using stun and this option didn't turn that on or off.

By storing the reflexive IP from stun as our local address, it broke
exit node when used with the `--stun` option.

Closes #1574

Signed-off-by: Russell Bryant <rbryant@redhat.com>