e2e: Add test coverage for a remote CIDR through a relay
This change adjusts an existing test, TestHubOrganization, and renames
it to TestHubAndSpoke.

Hub organizations were a concept a long time ago, but no longer exist.
Since that removal, this test no longer tested anything different than
TestBasicConnectivity.

By changing nodes to use `--relay-only`, we restore this test to
ensure connectivity works when deployed in a hub-and-spoke
architecture.

The test also included an intended scenario where a node tried to
reach an advertised CIDR from the remote node. Unfortunately, due a
typo, a node was pinging itself instead. This change makes the
intended test occur.

I was able verify that this test fails prior to my changes included in
this same PR.

Signed-off-by: Russell Bryant <rbryant@redhat.com>