ui: Don’t show the Users menu item.
Since users don’t have access to view list other users, it does not make sense to show the Users resource page in the UI, it’s only going to list the currently logged in user.  Maybe we bring it back if we have a Super Admin type role that
can view users across different tennants.

Signed-off-by: Hiram Chirino <hiram@hiramchirino.com>