nexd: reuse the wireguard listen port on restart
This issue reported that when a device is restarted, it triggered
updates to be sent to all peers. The reason updates were being sent is
that the Endpoints changed on every restart. When nexd started, it
would allocate a new random UDP listen port. This is bad, because it
means the data plane is interrupted at every restart until peering is
re-established using the new endpoint details.

This change uses the state store to save off the random UDP port so it
can be used on a restart.

There is one other behavior change included here. Previously, for a
relay, the default wireguard port was always used for a relay. Now if
a port is specified via a command line argument, it will be used for
regular devices and relays.

Closes #958

Signed-off-by: Russell Bryant <rbryant@redhat.com>