Write integration test for derp relay integration
Integration tests varify following scenarios
*.* Connectivity failure between peers behind symmetric NAT
in the absense of derp relay.
*.* Connectivity success between peers behind symmetric NAT
in the presence of public relay OR Onboarded Relay
*.* Connectivity success between peers where one peer is
behind symmetric NAT but the other one is behind reflexive
address, in the presence of public relay or Onboard Relay.
*.* Connectivity secnario where public relay goes away and
new relay onboarded by user. Test checks that the connectivity
fails when public relay is stopped, but it establishes back
one the relay is onboarded.

It also fixes :
- Update github workflow to run the derp relay test.
- Fix unit tests given that we are assuming that one public
DERP relay server will be available always.

Signed-off-by: Anil Kumar Vishnoi <avishnoi@redhat.com>